support online depth fix and depth remove (including export)

+++ src/render/ColorDepthFixer.h
@@ -3,6 +3,6 @@
 class cColorDepthFixer
 {
 public:
-    static cv::Mat FixColorDepth(const cv::Mat &color, const cv::Mat &depth, double alpha = 1);
-    static void FixColorDepth(std::string color_path, std::string depth_path, double alpha = 1);
+    static cv::Mat FixColorDepth(const cv::Mat &color, const cv::Mat &depth, float alpha = 1);
+    static void FixColorDepth(std::string color_path, std::string depth_path, float alpha = 1);
 };
